라벨 삭제 수정 - label null이 아닐 경우 삭제 조건 추가

+++ src/main/java/com/example/todoapp/repository/LabelRepository.java
@@ -44,6 +44,8 @@ public class LabelRepository {
 
     public void deleteLabel(Long label_id){
         Label label = em.find(Label.class, label_id);
-        em.remove(label);
+        if(label != null){
+            em.remove(label);
+        }
     }
 }
